complete migration to mui

+++ src/redux/contacts/operations.js
@@ -1,9 +1,19 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
-axios.defaults.baseURL = 'https://66bb51236a4ab5edd63803ff.mockapi.io';
+import { setAuthHeader } from '../auth/operations';
+axios.defaults.baseURL = 'https://connections-api.goit.global';
 
 export const fetchContacts = createAsyncThunk('contacts/fetchAll', async (_, thunkAPI) => {
   try {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (token === null) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthHeader(token);
+
     const response = await axios.get(`/contacts`);
     return response.data;
   } catch (error) {
@@ -13,6 +23,15 @@ export const fetchContacts = createAsyncThunk('contacts/fetchAll', async (_, thu
 
 export const addContact = createAsyncThunk('contacts/addContact', async (newContact, thunkAPI) => {
   try {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (token === null) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthHeader(token);
+
     const response = await axios.post(`/contacts`, newContact);
     return response.data;
   } catch (error) {
@@ -22,6 +41,15 @@ export const addContact = createAsyncThunk('contacts/addContact', async (newCont
 
 export const deleteContact = createAsyncThunk('contacts/deleteContact', async (id, thunkAPI) => {
   try {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (token === null) {
+      return thunkAPI.rejectWithValue('No token found');
+    }
+
+    setAuthHeader(token);
+
     const response = await axios.delete(`/contacts/${id}`);
     return response.data;
   } catch (error) {
